refactor: isolate LLM service capability
