+++ backend/services/llm/__init__.py
@@ -1,0 +1,13 @@
+"""LLM clients, lifecycle management, and provider factories."""
+
+from .factory import create_llm_client, get_available_models, validate_model_config
+from .service import LLMService, get_llm_service, init_llm_service
+
+__all__ = [
+    "LLMService",
+    "create_llm_client",
+    "get_available_models",
+    "get_llm_service",
+    "init_llm_service",
+    "validate_model_config",
+]
